Refactor data layer by moving team-related entities, DAOs, and repositories to a dedicated `data.team` package. Remove unused `AbilityRecordEntity` and `AbilityNameEntity` from `TeamDatabase`. Update all dependent use cases and DI modules with the new package imports.
