+++ app/src/main/java/com/sedilant/yambol/UpdateTeamUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.sedilant.yambol
 
-import com.sedilant.yambol.data.TeamRepository
+import com.sedilant.yambol.data.team.TeamRepository
 import com.sedilant.yambol.domain.UpdateTeamUseCase
 import javax.inject.Inject
