+++ app/src/main/java/com/sedilant/yambol/data/queries/ObjectivesAndTeam.kt
@@ -2,8 +2,8 @@ package com.sedilant.yambol.data.queries
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import com.sedilant.yambol.data.entities.TeamEntity
-import com.sedilant.yambol.data.entities.TeamObjectivesEntity
+import com.sedilant.yambol.data.team.TeamEntity
+import com.sedilant.yambol.data.team.TeamObjectivesEntity
 
 data class ObjectivesAndTeam (
     @Embedded val objectivesEntity: TeamObjectivesEntity,
