+++ app/src/main/java/com/sedilant/yambol/data/queries/PlayerAndTeam.kt
@@ -2,8 +2,8 @@ package com.sedilant.yambol.data.queries
 
 import androidx.room.Embedded
 import androidx.room.Relation
-import com.sedilant.yambol.data.entities.PlayerEntity
-import com.sedilant.yambol.data.entities.TeamEntity
+import com.sedilant.yambol.data.team.PlayerEntity
+import com.sedilant.yambol.data.team.TeamEntity
 
 data class PlayerAndTeam(
     @Embedded val playerEntity: PlayerEntity,
