+++ app/src/main/java/com/sedilant/yambol/data/team/PlayerEntity.kt
@@ -1,4 +1,4 @@
-package com.sedilant.yambol.data.entities
+package com.sedilant.yambol.data.team
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
@@ -6,7 +6,8 @@ import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
 
-@Entity(tableName = "team_objectives",
+@Entity(
+    tableName = "player",
     foreignKeys = [
         ForeignKey(
             entity = TeamEntity::class,
@@ -15,12 +16,14 @@ import androidx.room.PrimaryKey
             onDelete = ForeignKey.CASCADE // O RESTRICT it depends
         )
     ],
-    indices = [Index("team_id")])
-data class TeamObjectivesEntity(
+    indices = [Index("team_id")]
+)
+data class PlayerEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
-    val description: String,
-    val isFinish: Boolean,
+    val name: String,
+    val number: Int,
+    val position: Int,
     @ColumnInfo(name = "team_id")
     val teamId: Int,
 )
