+++ app/src/main/java/com/sedilant/yambol/data/team/TeamEntity.kt
@@ -1,4 +1,4 @@
-package com.sedilant.yambol.data.entities
+package com.sedilant.yambol.data.team
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
