+++ app/src/main/java/com/sedilant/yambol/data/team/TeamObjectivesDao.kt
@@ -1,11 +1,10 @@
-package com.sedilant.yambol.data
+package com.sedilant.yambol.data.team
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Update
-import com.sedilant.yambol.data.entities.TeamObjectivesEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
