+++ app/src/main/java/com/sedilant/yambol/data/team/TrainCrossTrainTaskEntity.kt
@@ -1,4 +1,4 @@
-package com.sedilant.yambol.data.entities
+package com.sedilant.yambol.data.team
 
 import androidx.room.Entity
