+++ app/src/main/java/com/sedilant/yambol/di/DataModule.kt
@@ -5,15 +5,15 @@ import android.content.Context
 import com.firebase.ui.auth.AuthUI
 import com.google.firebase.auth.FirebaseAuth
 import com.sedilant.yambol.data.DataStoreManager
-import com.sedilant.yambol.data.PlayerDao
+import com.sedilant.yambol.data.team.PlayerDao
 import com.sedilant.yambol.data.StatsDao
 import com.sedilant.yambol.data.StatsRecordRepository
 import com.sedilant.yambol.data.StatsRecordRepositoryImpl
-import com.sedilant.yambol.data.TeamDatabase
-import com.sedilant.yambol.data.TeamObjectivesDao
-import com.sedilant.yambol.data.TeamRepository
-import com.sedilant.yambol.data.TeamRepositoryImpl
-import com.sedilant.yambol.data.TrainingDao
+import com.sedilant.yambol.data.team.TeamDatabase
+import com.sedilant.yambol.data.team.TeamObjectivesDao
+import com.sedilant.yambol.data.team.TeamRepository
+import com.sedilant.yambol.data.team.TeamRepositoryImpl
+import com.sedilant.yambol.data.team.TrainingDao
 import com.sedilant.yambol.data.draftTrain.TrainingDatabase
 import com.sedilant.yambol.data.draftTrain.TrainingDraftDao
 import com.sedilant.yambol.data.draftTrain.TrainingDraftRepository
