+++ app/src/main/java/com/sedilant/yambol/domain/CheckJerseyNumberUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.sedilant.yambol.domain
 
-import com.sedilant.yambol.data.TeamRepository
+import com.sedilant.yambol.data.team.TeamRepository
 import javax.inject.Inject
 
 class CheckJerseyNumberUseCaseImpl @Inject constructor(
