+++ app/src/main/java/com/sedilant/yambol/domain/get/GetPlayerByIdUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.sedilant.yambol.domain.get
 
-import com.sedilant.yambol.data.TeamRepository
+import com.sedilant.yambol.data.team.TeamRepository
 import com.sedilant.yambol.ui.home.models.PlayerUiModel
 import javax.inject.Inject
