+++ app/src/main/java/com/sedilant/yambol/domain/insert/InsertPlayerUseCaseImpl.kt
@@ -1,8 +1,7 @@
 package com.sedilant.yambol.domain.insert
 
-import com.sedilant.yambol.data.entities.PlayerEntity
-import com.sedilant.yambol.data.TeamRepository
-import com.sedilant.yambol.domain.Position
+import com.sedilant.yambol.data.team.PlayerEntity
+import com.sedilant.yambol.data.team.TeamRepository
 import com.sedilant.yambol.ui.home.models.PlayerUiModel
 import javax.inject.Inject
