+++ app/src/main/java/com/sedilant/yambol/domain/insert/InsertTeamUseCaseImpl.kt
@@ -1,7 +1,7 @@
 package com.sedilant.yambol.domain.insert
 
-import com.sedilant.yambol.data.entities.TeamEntity
-import com.sedilant.yambol.data.TeamRepository
+import com.sedilant.yambol.data.team.TeamEntity
+import com.sedilant.yambol.data.team.TeamRepository
 import javax.inject.Inject
 
 class InsertTeamUseCaseImpl @Inject constructor(
